chore(persist-repo-scoped-fixes): apply nax-finish review fixes
- Move persisted-repo-scoped-fix.test.ts content into the spec-named homes:
  round-trip + absent-field into prd-auto-default.test.ts (US-001 AC1/AC2);
  reset-invariant into prd-reset-failed.test.ts (US-001 AC3-AC6).
- Create the spec-named US-002 test files:
  test/unit/execution/story-orchestrator/repo-scoped-fix-record.test.ts (AC1-AC8);
  test/unit/pipeline/stages/execution-repo-scoped-fixes.test.ts (AC9-AC13).
- Consolidate _executionDeps stub boilerplate into withExecutionDeps helper.
- Type the persisted fix fixtures with PersistedRepoScopedFix.
- Replace inline rmSync with cleanupTempDir helper.
- Rename AC14 label to AC13b (spec only defines AC1-AC13).
